+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - GenerateExamples - Add title and operationId to example
 - GenerateExamples - Generate subscriptionId in guid format
 - ModelValidator - Add type validation for additionalProperties whose type is specified as an object
+- Bugfix: FileLoader - isUnderFileRoot function supports validation for absolute path
 - API Scenario
   - Add roleAssignment step
   - Bugfix: Use content-type specified in swagger