+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # Change Log - oav
 
 ## 10/20/2022 3.2.1
+
+- ModelValidator - Fix bug during reporting real exampleJsonPath
 - ModelValidator - Add LRO header validation for data-planes swagger
 - GenerateExamples - Add title and operationId to example
 - GenerateExamples - Generate subscriptionId in guid format